feat: Sync the code with `cargo-c`.
As discussed in
https://github.com/wasmerio/wasmer/issues/2429#issuecomment-864925634,
it would be ideal if this crate could get synchronized with `cargo-c`
which has a very similar code but which supports more targets.

That's exactly what this patch does. It changes the code from a
`if`+`else` workflow to a `match` version. It also adds support for
`android`, `freebsd`, `dragonfly`, `netbsd` and `ios`.